RE-41:
