+++ internal/pay-service/entity/payment.go
@@ -10,3 +10,13 @@ type Payment struct {
 	Status    int       `json:"status"`
 	Balance   float64   `json:"balance"`
 }
+
+type Transaction struct {
+	ID            int       `json:"id"`
+	TransactionID string    `json:"transactionId"`
+	UserID        int       `json:"user_id"`
+	Amount        int64     `json:"amount"`
+	Type          string    `json:"type"`
+	Status        string    `json:"status"`
+	CreatedAt     time.Time `json:"created_at"`
+}
